requirements.txt
flask
flask-cors
python-dotenv
sentence-transformers
PyPDF2
groq
pinecone-client
werkzeug
